server/requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==4.0.1
blinker==1.5
cachelib==0.10.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.46
urllib3==1.26.13
Werkzeug==2.2.2
